+++ CHANGELOG.md
@@ -39,6 +39,7 @@ tech changes will usually be stripped from release notes for the public
 -   Spell tool: right-click was opening the context-menu
 -   Spell tool: once used, the select tool was not being properly activated
 -   Spell tool: the 'is public' label was wrongly attached to the range input box
+-   Fake Player: showing DM layer
 
 ## [2022.3.0] - 2022-12-12